Create worker prcess which spawns jobs and sends DeferredResult
Worker listens on the socket for incommming connections from the client.
When client is connected it transfers a pickled instance of RunnableTask
which is passed to a new Job instance.

Job instance can manage given tasks and execute them asynchronously.
It also keeps track of completed tasks and sends signal on completion.