Introduce SettingsProvider loading and managing settings.
Create the SettingsProvider class that loads configuration from a yaml
file and stores it.
SettingsProvider is instantiated empty and it can be populated with
values from the file, dictionary or manually. Configuration for each
service is wrapped in ServiceConfigurationProvider objects that manage
data extraction and validation; contrary to the previous solution,
where configuration files were parsed and validated by FormFactory
and Scheduler. Fix #34 Fix #28
Old config module is removed.

Add Singleton metaclass which ensures that only a single instance of
the class can be created. It overrides the `__call__` method of the
class and stores cached singleton objects for each class.

Make SettingsProvider and FormFactory singletons.

Forms are now stored and initialized in the FormFactory singleton
rather than being dynamically added to the module.

Split project creation into smaller components for better
clarity.