Create Runner - interface for job execution
Separated the configuration and runner selection from the scheduler to a
separate module. New RunnerFactory keeps and manages the configurations
and limits.

Executors are replaced by Runners which are instantiated by the factory
and passed the configuration based on the associated limits. Each runner
instance corresponds to a single run rather than the configuration.
They do not only launch the job, but also perform command construction,
working directory creation and creating file links.
Abstract Runner needs to be subclassed implementing its `submit` method.
Each runner must also provide a JobHandler interface for watching the
job status.

Limits are moved to a new module and metaclass LimitsMeta that checks
project specific class definitions is added. LimitsMeta makes sure that
the limits classes are properly constructed and raises an error on class
creation. It allows to spot errors immediately on startup.

New JobHandler acts as an interface and its construction is fully
controlled by the runner and handler themselves. It should provide
serialize and deserialize methods allowing to store it in the database
and recreate it on system restart.