Fix directory paths configured by the config loader.
Home paths provided in the -h parameter or SLIVKA_HOME were used
as is without resolving symlinks. On the other hand, os.getcwd
and paths to input and output files were formed with os.realpath
which resolved all symlinks.
It was actually a problem with an api code converting paths to
ids in a hacky way, but changing all paths to real paths works too.
Fix #129